feat: add entities endpoint
Issues:
closes #142
